feat: add reading session duration tracking
Track time spent reading each Surah and log sessions >= 10s
to a local Room table (reading_sessions). Matches web app's
logReadingSession store action. Sessions auto-pruned to 500.

+++ app/src/main/java/com/nur/quran/ui/viewmodels/PlannerViewModel.kt
@@ -716,7 +716,7 @@ class PlannerViewModel @Inject constructor(
         planDeltas[dayNumber] = total
         viewModelScope.launch(kotlinx.coroutines.Dispatchers.IO) {
             try {
-                repository.logReadingSession(delta, "reading", chapterId)
+                repository.logReadingSession(delta.toInt(), "reading", chapterId)
             } catch (_: Exception) { /* planner time stays invisible rather than crash */ }
         }
     }
